:recycle: Feat(client): Breaking update client api to be Tevm
