+++ extensions/viem/src/ViemTevmClient.ts
@@ -1,5 +1,5 @@
 import type { Tevm } from '@tevm/api'
 
 export type ViemTevmClient = {
-	tevm: Omit<Tevm, 'request'>
+	tevm: Tevm
 }
