+++ vm/client/docs/API.md
@@ -6,15 +6,9 @@
 
 ## Exports
 
-### Type Aliases
-
-- [MemoryClient](type-aliases/MemoryClient.md)
-- [RemoteClient](type-aliases/RemoteClient.md)
-
 ### Functions
 
-- [createMemoryClient](functions/createMemoryClient.md)
-- [createRemoteClient](functions/createRemoteClient.md)
+- [createClient](functions/createClient.md)
 
 ***
 Generated using [typedoc-plugin-markdown](https://www.npmjs.com/package/typedoc-plugin-markdown) and [TypeDoc](https://typedoc.org/)
