--- vm/client/docs/functions/createMemoryClient.md
@@ -1,20 +1,0 @@
-**@tevm/client** ∙ [README](../README.md) ∙ [API](../API.md)
-
-***
-
-[API](../API.md) > createMemoryClient
-
-# Function: createMemoryClient()
-
-> **createMemoryClient**(`params`): `Promise`\<`Client`\<`HttpTransport`, `undefined`, `undefined`, `PublicRpcSchema`, `object` & `PublicActions`\<`HttpTransport`, `undefined`\>\>\>
-
-## Parameters
-
-▪ **params**: `CreateEVMOptions` & `object`
-
-## Source
-
-[vm/client/src/createMemoryClient.js:7](https://github.com/evmts/tevm-monorepo/blob/main/vm/client/src/createMemoryClient.js#L7)
-
-***
-Generated using [typedoc-plugin-markdown](https://www.npmjs.com/package/typedoc-plugin-markdown) and [TypeDoc](https://typedoc.org/)
