--- vm/client/docs/functions/createRemoteClient.md
@@ -1,22 +1,0 @@
-**@tevm/client** ∙ [README](../README.md) ∙ [API](../API.md)
-
-***
-
-[API](../API.md) > createRemoteClient
-
-# Function: createRemoteClient()
-
-> **createRemoteClient**(`params`): `Client`\<`HttpTransport`, `undefined`, `undefined`, `PublicRpcSchema`, `object` & `PublicActions`\<`HttpTransport`, `undefined`\>\>
-
-## Parameters
-
-▪ **params**: `object`
-
-▪ **params.url**: `string`
-
-## Source
-
-[vm/client/src/createRemoteClient.js:7](https://github.com/evmts/tevm-monorepo/blob/main/vm/client/src/createRemoteClient.js#L7)
-
-***
-Generated using [typedoc-plugin-markdown](https://www.npmjs.com/package/typedoc-plugin-markdown) and [TypeDoc](https://typedoc.org/)
