--- vm/client/docs/type-aliases/MemoryClient.md
@@ -1,21 +1,0 @@
-**@tevm/client** ∙ [README](../README.md) ∙ [API](../API.md)
-
-***
-
-[API](../API.md) > MemoryClient
-
-# Type alias: MemoryClient`<>`
-
-> **MemoryClient**\<\>: `Awaited`\<`ReturnType`\<*typeof* [`createMemoryClient`](../functions/createMemoryClient.md)\>\>
-
-## Type parameters
-
-| Parameter |
-| :------ |
-
-## Source
-
-[vm/client/src/createMemoryClient.js:19](https://github.com/evmts/tevm-monorepo/blob/main/vm/client/src/createMemoryClient.js#L19)
-
-***
-Generated using [typedoc-plugin-markdown](https://www.npmjs.com/package/typedoc-plugin-markdown) and [TypeDoc](https://typedoc.org/)
