--- vm/client/docs/type-aliases/RemoteClient.md
@@ -1,21 +1,0 @@
-**@tevm/client** ∙ [README](../README.md) ∙ [API](../API.md)
-
-***
-
-[API](../API.md) > RemoteClient
-
-# Type alias: RemoteClient`<>`
-
-> **RemoteClient**\<\>: `ReturnType`\<*typeof* [`createRemoteClient`](../functions/createRemoteClient.md)\>
-
-## Type parameters
-
-| Parameter |
-| :------ |
-
-## Source
-
-[vm/client/src/createRemoteClient.js:16](https://github.com/evmts/tevm-monorepo/blob/main/vm/client/src/createRemoteClient.js#L16)
-
-***
-Generated using [typedoc-plugin-markdown](https://www.npmjs.com/package/typedoc-plugin-markdown) and [TypeDoc](https://typedoc.org/)
