+++ vm/client/package.json
@@ -68,6 +68,7 @@
 	},
 	"devDependencies": {
 		"@ethereumjs/util": "^9.0.1",
+		"@tevm/api": "workspace:1.0.0-next.22",
 		"@tevm/server": "workspace:^",
 		"@tevm/tsconfig": "workspace:^",
 		"@tevm/tsupconfig": "workspace:^"
