+++ vm/client/src/index.ts
@@ -1,2 +1,1 @@
-export { createRemoteClient, type RemoteClient } from './createRemoteClient.js'
-export { createMemoryClient, type MemoryClient } from './createMemoryClient.js'
+export { createClient } from './createClient.js'
